feat: migrate application to use PostgreSQL
Updated the application to support PostgreSQL as the database.
This includes modifying the connection strings in the configuration
files, updating the database context in DependencyInjection, and
removing SQL Server-specific code from migration files.

Additionally, new JWT configurations were added, and the
`app.ApplyMigrations();` method was included in the
development environment to ensure migrations are applied at
startup. The column types for certain properties were also
adjusted to use PostgreSQL-specific types. A new
`MigrationExtensions.cs` file was created to encapsulate
migration logic, and a new migration file was added to define
the initial schema for PostgreSQL.