chore: update dotnet version and build command
The code changes in the `publish-workflow.yml` file involve
updating the .NET version used in the workflow from a generic
setup to a specific solution file. The line that previously
built the project without specifying a solution file has been
replaced with a command that builds the `fiap-cloud-games.sln`
solution file directly. This change ensures that the build
process targets the correct solution, potentially improving
build accuracy and efficiency.