chore: update Dockerfile path in publish workflow
The publish-workflow.yml file has been modified to remove the
Dockerfile path for the Refit.Api project and add the path for
the FIAPCloudGames.Api project. This change reflects a shift in
the focus of the build process to the FIAPCloudGames.Api
project, ensuring that the correct Dockerfile is used during
the build and push process.