chore: update Dockerfile path for ACR build process
The Dockerfile path used in the build and push process for the
container image to Azure Container Registry (ACR) has been
updated. The original path (`FIAPCloudGames.Api/Dockerfile`)
has been replaced with a new path (`src/FIAPCloudGames.Api/Dockerfile`).
This change reflects a modification in the directory structure
where the Dockerfile is located.